feat: Working product
Working product with timer, nether portal generation and loot on start

Issue: #1
